requirements.txt
streamlit
openai
tavily-python